Seperate hidden and public cogs.
